M1: PC scaffold, stub shim, headless file-I/O proof

+++ source/system/gte.h
@@ -52,13 +52,13 @@
 
 	/* Portable equivalents: data regs 9,10,11 = IR1,IR2,IR3 */
 
-	#define gte_stsv_x( r0 )	(*(short*)(r0)=(short)GTEport_GetData(9))
-	#define gte_stsv_y( r0 )	(*(short*)(r0)=(short)GTEport_GetData(10))
-	#define gte_stsv_z( r0 )	(*(short*)(r0)=(short)GTEport_GetData(11))
+	#define gte_stsv_x( r0 )	(*(s16*)(r0)=(s16)GTEport_GetData(9))
+	#define gte_stsv_y( r0 )	(*(s16*)(r0)=(s16)GTEport_GetData(10))
+	#define gte_stsv_z( r0 )	(*(s16*)(r0)=(s16)GTEport_GetData(11))
 
-	#define gte_stlvl_x( r0 )	(*(long*)(r0)=(long)GTEport_GetData(9))
-	#define gte_stlvl_y( r0 )	(*(long*)(r0)=(long)GTEport_GetData(10))
-	#define gte_stlvl_z( r0 )	(*(long*)(r0)=(long)GTEport_GetData(11))
+	#define gte_stlvl_x( r0 )	(*(s32*)(r0)=(s32)GTEport_GetData(9))
+	#define gte_stlvl_y( r0 )	(*(s32*)(r0)=(s32)GTEport_GetData(10))
+	#define gte_stlvl_z( r0 )	(*(s32*)(r0)=(s32)GTEport_GetData(11))
 
 #endif	/* PSX_MIPS_ASM */
